# mindscape/models/responses/LoginResponse.go
/* Generated by egg v0.0.1 */

package responses

import (
	"github.com/adamkali/mindscape/db/repository"
	"github.com/labstack/echo/v4"
)

type LoginResponse struct {
	Data    *UserData `json:"data"`
	JWT     string    `json:"jwt"`
	Success bool      `json:"success"`
	Message string    `json:"message"`
} // @name LoginResponse

func NewLoginResponse() *LoginResponse {
	return &LoginResponse{Success: false, Message: ""}
}

func (LoginResponse *LoginResponse) Fail(ctx echo.Context, code int, err error) error {
	LoginResponse.Message = err.Error()
	return ctx.JSON(code, LoginResponse)
}

func (LoginResponse *LoginResponse) Successful(ctx echo.Context, user *repository.User, token string) error {
	LoginResponse.Data = UserDataFromRepository(user)
	LoginResponse.JWT = token
	LoginResponse.Success = true
	return ctx.JSON(200, LoginResponse)
}

func (LoginResponse *LoginResponse) Handle(
	ctx echo.Context,
	user *repository.User,
	code int,
	token string,
	err error,
) error {
	if err != nil {
		return err
	}
	return ctx.Redirect(200, "/dashboard/"+user.ID.String())
}