# mindscape/models/responses/StringResponse.go
/* Generated by egg v0.0.1 */

package responses

import (
	"github.com/labstack/echo/v4"
)

type StringResponse struct {
	Data    *string `json:"data"`
	Success bool    `json:"success"`
	Message string  `json:"message"`
} // @name StringResponse

func NewStringResponse() *StringResponse {
	return &StringResponse{Success: false, Message: ""}
}

func (StringResponse *StringResponse) Fail(ctx echo.Context, code int, err error) error {
	StringResponse.Message = err.Error()
	return ctx.JSON(code, StringResponse)
}

func (StringResponse *StringResponse) Successful(ctx echo.Context, stringLike string) error {
	StringResponse.Data = &stringLike
	StringResponse.Success = true
	return ctx.JSON(200, StringResponse)
}